+++ pruebas/MemoriaVisoespaciaPrueba.py
