+++ vistas/MVCWindowWidget.py
